(#1021) use browserify to insert globals, but working
