+++ package.json
@@ -57,8 +57,9 @@
   ],
   "scripts": {
     "jshint": "jshint -c .jshintrc src/*.js src/adapters/*.js",
-    "build-js": "mkdir -p dist && browserify src/pouch.js -o dist/pouchdb-nightly.js",
-    "watch-js": "mkdir -p dist && watchify src/pouch.js -o dist/pouchdb-nightly.js",
+    "build-js": "browserify -s Pouch -e src/pouch.js -o dist/pouchdb-nightly.js",
+    "build-js": "mkdir -p dist && browserify src/pouch.js -s Pouch -o dist/pouchdb-nightly.js",
+    "watch-js": "mkdir -p dist && watchify src/pouch.js -s Pouch -o dist/pouchdb-nightly.js",
     "uglify": "uglifyjs dist/pouchdb-nightly.js -mc > dist/pouchdb-nightly.min.js",
     "build": "npm run build-js && npm run uglify",
     "test-node": "./bin/test-node.js",
