+++ tests/pouch.shim.js
@@ -1,0 +1,1 @@
+global.Pouch = {};
